Revert "KSM: Start KSM by default"
This is not the standard behavior of KSM, intended to be started
by userspace daemons, but the idea may be fine.
However, the real reason for this revert is that the commit is
actually impacting on performance of embedded systems with slow
CPUs, like x86 Atom and, generally, ARM.

The actual performance impact is on CPU usage of KSM, as I can
see a 12% constant CPU usage by its daemon while the system is
processing heavy workloads.
We obviously need maximum CPU time in those conditions, thing
that KSM won't guarantee.

This reverts commit af91c534055a87f10de5f49a09525eaca8f2de92.

Change-Id: I07c11daba03341663ccdf6c9913898fb3165fb10